cc: Improve type annotation of Heap key function
Previously, the key function operated on Any by accident, which
suppressed potential issues from mypy. While the new annotation
is still not very strict (see note in comment), it is a lot more
specific than the old code was.

I had to move the Comparable type variable to the 'utils' module
so Heap could use it as well.

The 'total_ordering' decorator didn't actually add the remaining
comparator operators to the protocol, so I added them manually
instead.