cc: Clean up permission check of GetFactoryInfo
Instead of checking whether the user has all 'list' privileges,
we now check whether they have privileges to list all databases
for which information will be returned.

In practice there is no difference, but it's more robust to have
the permission check mimic the data access. Ideally we'd tie the
check and the access together, but that requires a major redesign.