cc: Remove status models
This was a prototype for a generic data structure that could be
fetched and observed via API calls. It was used for the notifier
client, a system tray application to monitor the last run result
of a number of selected schedules.

I think that a system tray application is probably no longer the
best way to monitor the status of schedules. Instead, we could
use web push notifications, chat integration (IRC, Slack,
Mattermost etc) and widgets that can be embedded in other sites.

For the server side, adopting GraphQL would be better than having
our own design.

But until I have time to pick up those tasks, I think we're better
off without the status model code in Git.