cc: Complete type annotations of ResourceControl page
When an anonymous user suspends or resumes a resource, the string
"anonymous" is now stored as the name of the user that made the
last change, rather than storing None. While storing None is probably
harmless in practice, it violates the type annotations of ResourceBase
and may not be harmless in the future.

A different solution would be to remove the 'changeduser' key from
the ResourceBase properties if an anonymous user made the change,
but we already have "ControlCenter" as a possible value, so having
a visible name is more consistent.