cc: Rename schedule "sequence" to "repeat"
Some places, such as the ScheduleDetails page and the ScheduleRepeat
enum, were already calling it "repeat", but most of the UI and code
were calling it "sequence". I think "repeat" is more specific and
therefore easier to understand.

I also added a property to fetch a schedule's repeat, so we don't
have to use __getitem__() all the time.