cc: Simply overwrite stale PID files
Previously there was a lot of defensive code that would try to avoid
starting a Control Center twice. However, that functionality should
be part of the init system and not in every individual service in
a half-baked fashion.

The risk of being too defensive is that if a server has an unclean
shutdown, there could be a PID file that contains an existing PID
that after the restart is claimed by different process. In that case,
we would refuse to start the Control Center and admin intervention
would be required to get it back up and running.