inv: Add option for detailed test case selection
We already had suite selection, which is useful if you want to do
a regression test before commit. But if a test fails, being able
to run only a single test case at a time is very useful.

The '-k' option of pytest can actually do much more than select
a single test, so we just pass the '--select' value as-is to
inherit those features.