cc: Remove support for on-demand loading of DB records
Instead, we explicitly preload all records.

The only table that wasn't already preloaded was products. While
that is a relatively large table, it's in the same order as task
runs, so I don't expect memory use to become unmanageable.

In the long term, both task runs and products should be stored in
a real database. Removing on-demand loading simplifies the code of
our custom database-like solution, which should make migration
easier.

The direct trigger to remove this support though was that the
key/value/item views used the undocumented and private '_mapping'
field of their superclasses. We only needed complex custom views
because of the load-on-demand mechanism.

A positive aspect of explicit preloading is that we no longer
load and/or create database records when modules are imported,
except for projectlib. This should make it easier to move towards
parallel unit test execution in the future.