cc: Redirect to relative URLs
This wasn't allowed by RFC-2616, but RFC-7231 obsoletes that one and
does aloow relative URLs in the Location header.

The problem with absolute URLs is that Twisted constructs a base URL
which may not match the externally visible URL when we're behind
a reverse proxy. Also, when not listening to a TCP socket, the
construction will fail to get a host and breaks.