cc: Always show execution graph on Design page
For something so core to SoftFab, it was a pity it was at 3 clicks
distance from the home page. Now it's at 2 clicks.

The reason it wasn't enabled by default is performance: GraphViz
needs some time to layout the graph and that can be felt in the page
loading time. Even on my modern PC, it takes 200 ms to load the Design
page, while many other pages load in under 10 ms.

There are things we can do about this, such as putting the SVGs in
image tags so they can be loaded after the page itself and caching
the generated graphs.