cc: Apply our own product filter instead of querying the graph
When drawing graphs for the ProductDetails page, we only want to see
the highlighted product as an input or output, so other products
should be filtered out. Previously, this was done by introspecting
the graph that was being built, but that is a feature of pyGraphViz
that not all alternative libraries have.