cc: Replace 'pygraphviz' by 'graphviz' library
Installing 'pygraphviz' with pip can be challenging, since it needs
to compile native code. I assumed that native code would allow it
to layout graphs without spawning processes, but the native code is
only used to create a data structure that is still processed by
'dot' running in an external process.