cc: Avoid duplicate XML IDs in Graphviz SVG output
When an ID attribute is set for the graph, that ID is prefixed to
all edge and node IDs. If the same string is passed as a name,
there is no prefix and there will be duplicate IDs on pages that
contain multiple inline graphs, such as the documentation page
about execution graphs.