cc: Clean up string encoding and decoding in 'graphview' module
The default encoding is 'utf-8', so we don't need to pass that.

The Graph.export() method was changed to always output in bytes:
previously, the output for the DOT format was a string. That string
would just be converted to bytes in Response.write(), so doing it
early is equally efficient while it simplifies the interface.