cc: Feed XML parser streaming data from SVG rendering
Instead of parsing the XML all at once, we now parse each piece of data
as its read from the 'dot' process. This is slightly more efficient,
since we don't have to buffer the data and we can act upon it while
it is still likely to be in the CPU cache. But mainly it gives us
more options to customize the parsing in the future.