cc: Fix presentation of unorderedList widget
This broke in 6532c881 (cc: Merge instead of overwrite 'class'
attribute on hgroup/vgroup), because I overlooked the fact that
_Group is a base class for not just hgroup/vgroup but also for
unorderedList.

APE didn't catch the problem in CI because we currently only run APE
on a Control Center with an empty database; we need to also run it
on a populated Control Center.

One of our PyLint plugins did catch the problem, but because we have
so many unresolved errors reported by PyLint, that went unnoticed.