cc: Revert incorrect fix in returnToReactor()
Yielding something other than a Deferred will cause @inlineCallbacks
to immediately call the wrapped function again. This means that
returnToReactor() was actually not passing control back to the
reactor.

By the way, ticket 448 existed in an internal Trac instance at
Philips, long before SoftFab went open source. It stated that
the GetJobHistory API call was raising an unhandled exception,
but that is no longer the case after this revert, so I'm assuming
the actual cause got fixed in the mean time.