cc: Fix handling of ConnectionDone
Until now, closed connections were only handled if ConnectionLost
was raised, which indicates a non-clean closing of the connection.
If a connection was closed cleanly, such as the user pressing the
stop button in a web browser, ConnectionDone is raised instead.

ConnectionLost was also not handled correctly: when notifyFinish()
has fired (which happens on ConnectionClosed), we should not call
finish() ourselves anymore.

While at it, I also changed the logging to use the Python logging
framework instead of Twisted's. This was picked because all our
other code logs in that way as well; I haven't evaluated the
merits of each logging framework yet.