cc: Encode Content-Disposition header according to RFC 6266
The original encoding routine was written many years ago to work
around limitations of browsers at that time: only Firefox and Opera
accepted RFC 2184 style necoding, for other browsers we had to use
browser-specific fallbacks.

Now we have RFC 6266 which not only seems to be supported better,
it also provides advice on interoperability with non-compliant
user agents.

I added a unit test as well. While it is very minimal, it acts
both as a sanity check and as a place where we can add additional
checks, should the need arise.

After implementing this, I discovered that we don't actually have
any file export at the moment that requires encoding Unicode or
difficult ASCII characters in the Content-Disposition header.
But the code is there in case we relax some naming rules in the
future or add new kinds of export.