cc: Avoid infinite redirect loop on widgets
Before this, requesting /Home/recentJobsTable on a private
factory would redirect to /Home/Login, which would again
redirect to /Home/Login, creating a loop. Eventually the
browser will give up, but it's not very efficient.

The solution is to make widgets (subpath URLs) fail immediately
when the user is not authenticated and anonymous guest access
is turned off.