cc: Replace the 'upgrade' module by the 'migrate' subcommand
The module contained a script that should be run after an upgrade to
migrate the database to the latest schema.

The script was actually broken because the project record was not
explicitly preloaded yet. This became necessary after we reduced
the amount of import side effects, which auto-loading was one of.