cc: Make --debug a global option and log migrations
Debug options are useful for all commands, not just 'server'.

This commit applies debug logging to the 'migrate' command. It also
logs the migration itself, since that is useful information when
analyzing problems.

The 'init' command currently ignores the --debug option, but
I expect that to change in the future.