cc: Add type stubs for importlib.metadata and resources
These stubs are shipped as part of mypy's included copy of typeshed,
but they have a version check to prevent the interface from being
available on Python versions that don't have these libraries.

I copied the stubs from typeshed, added a header to document their
origin, disabled the version check and renamed them to function as
type stubs for the importlib_metadata and importlib_resources
backports.

The end result is that we can type check code using these modules
on any Python version we support.