cc: Show reply body if API call completes with an HTTP error status
The server will explain what actually went wrong in the reply body,
so we should show that to the user instead of a generic error.

Since HTTPDownloader doesn't provide the reply body for error status
codes, we had to switch to Agent as an HTTP client. This was a good
idea anyway, since HTTPDownloader is deprecated. Unfortunately,
HTTPDownloader is marked deprecated in a way that pydoctor's static
analysis does not understand, so you have to look in the source code
rather than the HTML docs to find this out.