cc: Add 'softfab user add' CLI command
This can be used to create the initial operator. Additional users
accounts can then be created either on the command line or on the
web interface.

What is still missing is the ability to set a password for newly
created user accounts. This can be done manually through Apache's
'htpasswd' tool, but that is of course not a proper solution.

What is also missing is related commands, such as deleting a user
account and changing an account's role.