cc: Add method to construct API URLs
This avoids duplication.

The reason to put the method in GlobalOptions is that at some point
we might want to support accessing a remote Control Center via the
command line client and in that case we'll probably have a global
option to provide the CC's URL.

I renamed the host from "dummy" to "cmdline", so it is clearer if it
turns up in the logging. By default the URL is not logged though.