cc: Add 'softfab user remove' CLI command
Even though we don't recommand removing user accounts, there can be
situations where it is the right course of action. For example the
operator might want to remove an account that was never used.
Another valid reason is when the user requested to be removed for
privacy reasons.

The command doesn't actually work yet because run_DELETE() and the
handling of the DELETE method are not implemented, but the command
line interface part of it does.