cc: Do not refresh jobs table on UserDetails page for non-existing user
A non-existing user leads to the OwnedJobsTable not being presented,
but since it's listed by iterWidgets() and iterDataTables(), refresh
script is generated for it nevertheless. When that script tries to
fetch table updates, page generation breaks (KeyError).

Omitting the widget from iterWidgets() and iterDataTables() will make
the subpage requests return 404 instead of 500 and while the 404 page
does not look great, that is at least the right status code.

While this commit fixes the problem in practice, the does not
address the underlying issue, which is that the refresh and subpage
mechanisms are difficult to use correctly.