cc: Improve response when logging in with a removed user
If a user still exists in the password file but the user record is
removed, we used to report "Internal error" in an attempt to not
help attackers. However, this message was unique in the code, so
it is not all that hard for someone to figure out what happened.

Now we treat login with a removed user the same as login with
a user that does not exist or an incorrect password.