ccmgr: Improve handling of virtualenv activation failure
The error message stated that creation of the venv failed, but what
happens in the script is activation, not creation.

Additionally, instead of checking for the presence of the file, we
should just try to source it and complain if it doesn't work. There
are other reasons activation can fail after all, for example not
having read permissions on the file or an error during the execution
of the activation script.