cc: Provide type arguments for all uses of Retriever
This uncovers incorrect use of retrievers. To avoid making too many
changes at once, I used typecasts to ignore those incorrect uses,
but they should be properly fixed later.

Note that incorrect here means that the runtime types don't match
the declared types: retrievers can return None which is not
comparable. However, since None is mapped to 'missing' when sorting,
there should be no runtime errors as a result of this. But this
situation is fragile.