cc: Eliminate use of global productDefDB from productDB
Instead of fetching the product definition on demand, we now fetch it
when creating the Product object (record creation or loading). Unlike
productDB, productDefDB is always in memory, so there should be no
significant change in performance or memory use.

ProductDB.create() will now raise KeyError if passed a non-existing
product name instead of storing a None key in the created product.
This is an improvement, since Product.getDef() doesn't expect the key
to ever be None.