cc: Pass ResTypeDB excplicitly to Config.isConsistent()
This removes an import of the global DB.

Note that Config.isConsistent() is not actually called anywhere in
the code, but it is called from the unit test. While the check it
does is useful, it should probably be re-done in the style of
iterInputConflicts(), so that the inconsistencies it finds can
be reported to the user.