cc: Fix chunked calling of startup sequence
While using 'succeed(None)' didn't raise any errors, it also didn't
accomplish what we wanted to do, which is returning control to the
reactor. Instead, a single send() on the coroutine completed it.

Since 'asyncio.sleep()' handles a zero delay as a special case,
it does not interact with the event loop, so it doesn't matter
that we're using Twisted's reactor instead.