cc: Move database mapping and factory injection to 'databases' module
The unit tests broke because the factory injection was not redone in
reloadDatabases(). Since the unit tests do not go through the startup
sequence of SoftFabRoot, the solution was to move factory injection
to the 'databases' module. But to make that possible, the creation of
the databases mapping had to be moved as well.