cc: Fix migration mechanism
The first problem was that databases were not preloaded and therefore
would appear empty when iterated through, so no records would get
migrated. This was probably introduced when removing the on-demand
loading in December 2019.

A second problem was that dependency injection was not done when
migrating, which meant productDB had no access to productDefDB.
This was introduced earlier this week.