cc: Load resourceDB after taskRunDB
The _setExecutionrun() method that handles parse events looks up
the given task ID in taskRunDB, which won't exist if taskRunDB
hasn't been loaded yet.

The symptom of this was that an "Execution run <ID> does not exist"
message was logged and the run ignored if a Task Runner was in the
running state when the Control Center was started.