cc: Create resultlessJobs and unfinishedJobs in startup()
These were created globally in the joblib module until now.

This is not just a cleanup to remove global variables: the fact they
were created in joblib meant that they also ran their initial query
at that time, before the jobDB was loaded. So this was broken since
last December, but since we don't typically have waiting jobs when
the Control Center is restarted we didn't notice.

The best solution is to write automated system tests, but we need
a proper API first to be able to remote control a Control Center.