cc: Handle resource being reserved by a non-existing task
The presentation of the ResourceIndex page would break when a resource
is reserved by a task that does not exist. This situation can only
occur when the database is inconsistent, for example when a running
job was removed through the file system.

The only way for the user to recover the resource in this situation
is to remove it and create it again, or remove the 'reserved'
attribute in the resource XML. Given that this is not something
that can occur in normal use, I think that is acceptable.