cc: Fix attaching and detaching of ExecutionObserver
In b448f120 (Only register ExecutionObserver for TRs that are in
ResourceDB), the attaching and detaching of the ExecutionObserver
 was moved from the __init__() and retired() methods on the record
to the add() and remove() methods on the DB, respectively.

While moving the attaching out of __init__() was a good idea for the
reasons outlined in that commit's comment, moving it to add() was
incorrect, since that means the observer won't be added for records
that are loaded from the database, or if an edit replaces an old
record with a new record object. The symptom of this is that the
task will be shown as running, but the Control Center thinking the
Task Runner it was assigned to should not be running it.

The proper solution is to use the _register() and _unregister()
methods that get called in all relevant situations and were actually
designed for performing observer (de)registrations.