cc: Simplify runCoroutine()
Chaining the 'result' variable isn't something we need in our usage
pattern. By removing that, we no longer need two separate routines.

Adding type arguments will alert us if a coroutine is passed that
does care about values being passed in and/or out.

Note that there is a subtle difference between the old and new runtime
behavior, in that there is no longer a return to the reactor before
the first processing step. I think the new behavior is better, since
it can only pause execution at the points where the coroutine yields.