cc: Move updating of record out of SelectableABC
This was the only functionality from DatabaseElem that SelectableABC
uses. So with that removed, it no longer needs to be a mixin.

Note that now the responsibility for calling update() rests on the
caller of updateTags(), which is consistent with setTag() that didn't
call _notify() to begin with. In the future, we could use a context
manager to wrap tag updates in a way that forces a commit to the DB
on live records.