cc: Make tag values case-sensitive
Having them case-insensitive made both the interface and implementation
of tagging a lot more complex. This commit already removes a lot of
lines, but more simplifications are possible in the future.

Another issue is that we want to store case-sensitive paths in tags,
such as the names of (Git) branches that trigger a schedule.

The original reason for making tag values case-sensitive is that users
might not remember the case of a tag and type it in the wrong case,
accidentally creating a new tag. But the way the UI uses tags is to
show existing tags, so it's unlikely users will be doing much typing
of existing tag values.