cc: Make ScheduleManager responsible for creating jobs
The trigger() and createJobs() methods were already separated,
but the error handling for createJobs() was still in trigger().

Note that we now update the 'last run' info if createJobs() failed.
While I think the new behavior is better, I don't expect this
situation to come up often (or at all, ideally) in practice.