cc: Make ScheduleManager responsible for schedule 'running' status
Previously, we let Scheduled.isRunning() query the last run's jobs
until the result was 'not running'. Now we instead keep track of
which jobs are still running in ScheduleManager.

The main advantage is that Scheduled no longer needs to access
jobDB, while ScheduleManager already had access to jobDB via
dependency injection.

This new approach is also slightly more efficient, since we don't
need to iterate over jobs to determine a schedule's status.

The change in the unit test was necessary because ScheduleManager
only tracks schedules that are in the DB, while the test assumed
that the 'running' status would be continued to be updated also
after the schedule was removed from the DB.