cc: Make SoftFabRoot responsible for populating the new-style API
While we currently have two separate roots, eventually the new-style
API will become part of the main CC root.

The motivation for doing this change now is that we can now populate
the new-style API after database loading, which is necessary when
we want to remove the global userDB.

It also avoids the API operating on incomplete data during the startup
phase. We should also serve a 503 page, but for now returning 404 is
better than working with incomplete data.