cc: Move RunInfo from taskrunlib to resourcelib module
In commit 56a29b2a (Move RunInfo from taskrunnerlib to taskrunlib
module), the RunInfo class was moved to taskrunlib, to get rid of
the circular import caused by importing jobDB from joblib.

Now that jobDB is passed as an argument, that motivation no longer
applies. Since RunInfo is only used as part of TaskRunnerData,
having it in resourcelib (which taskrunnerlib was merged into)
makes more sense.