cc: Pass project object to all pages as a dependency
Every UI page shows the project name in its title bar, so I decided
to put the object in the PageProcessor base class. Some API calls
might not need it, but this is a big time saver and still an
improvement over a global variable.

Maybe it would make more sense to have this as an instance variable
instead of a class variable on the processor, but I used class
variables elsewhere, so let's stick with that for now.