cc: Add a run to a task as soon as it is created
This eliminates the Job.newRun() method, which was intended for
supporting reruns at some point, but that is probably not something
we want to implement. And if we change our mind about that, it's
likely the DB approach will be different by then, so it's still not
worth keeping this extra complexity around.

Task.create() is renamed to Task.addTask() to be consistent with how
other factory methods are named: 'create' if they only create a record
in memory, 'add' if they also add it to a database.