cc: Remove on-demand fetching of Job and Task in TaskRun
This extra complexity isn't necessary anymore now that we preload
databases during migration.

When the job is not found, ObsoleteRecordError is now also raised
when not migrating. During normal startup, we don't handle neither
KeyError nor ObsoleteRecordError, so the result is an error being
logged in either case. Always raising ObsoleteRecordError is both
easier and leads to a more specific log message.