cc: Fix incorrect argument default of Framework.isFinal()
The parent of a framework is paramTop, the dummy object at the top
of every parameter inheritance chain. However, since this object is
returned unconditionally by Framework.getParent(), we don't need to
create a function that returns paramTop and can instead use None as
the default.

This eliminates the use of the global frameworkDB variable inside
the frameworklib module, which was the reason I was looking at that
code in the first place.