cc: Do not call _iterDatabases() twice when reloading
When this function was returning only global DB instances, it would
return the same objects if called twice. But now that is is also
constructing DB instances, calling it a second time will return
a different object. This means that the DB that will be preloaded
will not be the one that was used during dependency injection.