cc: Improve handling of empty JUnit test reports
Use result 'cancelled' and summary 'no test cases found'.

The previous result was None, meaning that no result could be extracted
from the report. Since generally JUnit XML does produce a result, it is
likely wrappers rely on this and have the JUnit XML as the only result
source. In that case, an empty report would lead to no result at all
for a task, which SoftFab considers an error.

The choice for 'cancelled' was motivated by the assumption that the
most likely cause is not an empty test suite but instead all tests
being deselected. This is similar to all tests being skipped and
'cancelled' is the result code we use for skips.

The previous summary was '0 failed' and while technically correct,
it did not draw attention to the fact that nothing was executed.