cc: Ignore skipped test cases when determining task result
If combined like task results in a job, a test task with only passed
and skipped tests would end up with result 'cancelled', which is
likely not what the user wants to see. While tasks getting cancelled
is typically unexpected, skipped test cases are typically marked as
such and therefore unsurprising, which means we don't need to draw
attention to them by altering the combined result.