cc: Disable new raise-missing-from pylint rule
This rule was added in pylint 2.6. I checked a few of the issues it
flagged, but they were mostly trivial conversions from KeyError raised
by a 1-line try body to a more specific exception, so chaining
exceptions would not make the backtraces more useful.

Perhaps it's worth revisiting the issues it flagged some day, but in
practice it's also fine to add a 'from' clause while debugging and
encountering an exception that is not easy to trace to its root.