cc: Allow guest users to change their own password
In the past we used shared guest accounts, where it is not desirable
for one guest to change the password. However, shared accounts are
a bad idea for security, since it's difficult to track who has access
and revoking access for a single person requires communicating a new
password to all the remaining users of the account.

Also in the future we want notifications to be configured on
a per-account basis, such that guests and other users can enable
notifications for parts of the factory that they read access to,
even if they don't have write access. Notifications are impractical
with shared accounts, since it there wouldn't be a single destination
for the notification messages.

When accounts are personal (non-shared), there is no reason that
the account owner should not be allowed to change the password.
Additionally, removing that special case will simplify the
password reset workflow.