cc: Add 'softfab user block' CLI command
Blocking a user sets their role to inactive (which prevents new
logins and ends existing sessions) and removes their password.

This command is intended to deactivate a user account when a user
leaves a project. Since many people have the unfortunate habit of
reusing passwords across sites, the responsible thing to do is
remove the password when it is no longer needed.

Another reason to remove the password is that the password file
is in a standard format such that for example a web server can
be configured to use it as well. In a scenario like that, a blocked
user should no longer have access to the other resources, which
can only be achieved by removing or invalidating their entry in
the password file.