cc: Make typing_extensions a runtime instead of developer requirement
I want to be able to use its runtime type inspection features as well.

The reason for importing the type extensions from the softfab.compat
module was to provide dummy values at runtime. Now the need for that
is gone and we can simplify things by importing from typing_extensions
directly.