cc: Generate reference document for CLI syntax
This document page is generated from the Click objects and their help
strings, so it should automatically stay in sync when the CLI is
updated. However, Click has a very rich interface and I've only
implemented support for the parts of it that we currently use,
so some CLI updates might require some new doc rendering code
as well.