cc: Re-add CLI check whether 'softfab.ini' exists in data dir
The check is now performed when requesting an agent, since agents
are only used for executing commands that perform API calls and
therefore need an initialized data directory.

I added an additional check whether the control socket exists,
informing the user of the most likely cause of it not existing,
that being that the Control Center is not running.