cc: Make HTML/XML processing instruction (PI) handling more flexible
Instead of taking a mapping of handlers, the parser now takes a single
callable. This makes it easier to implement in a more dynamic way.

An additional advantage is that by making PIHandler a Protocol,
we simplify the documentation and type annotations, as we now have
a "home" location to attach those to.