cc: Remove obfuscation blur over webhook secret
Neither GitHub nor Gogs obfuscates the secret field, so there is
little point in trying it obfuscate it on our end.

The way it was supposed to work is that the user can select and
copy-paste the obfuscated text, but since the selection itself is
also blurred, this is not a good user experience.